+++ bootique-linkmove4-csv/src/test/java/io/bootique/linkmove/v4/csv/cayenne/Table1.java
@@ -1,0 +1,9 @@
+package io.bootique.linkmove.v4.csv.cayenne;
+
+import io.bootique.linkmove.v4.csv.cayenne.auto._Table1;
+
+public class Table1 extends _Table1 {
+
+    private static final long serialVersionUID = 1L;
+
+}